test(core-domain): share EpubHref tests to commonTest for iOS uriPath parity (#946)
normalizeEpubHref moved to commonMain with a platform uriPath seam
(java.net.URI on JVM, NSURL on iOS). Consolidate the existing jvmTest
EpubHrefTest into commonTest and add normalizeEpubHref + epubCfiToSpineIndex
cases so the two actuals are proven to agree on-device (11 tests, iOS + JVM).
Wire :core:domain:iosSimulatorArm64Test into the iOS commonTest CI job.

One resolveEpubHref test name contained a comma (illegal in Kotlin/Native
identifiers) and was renamed; behaviour unchanged.

Removed-test: resolves against a base folder, preserving the fragment
